Moved classes to yii\data namespace, adjusted guide
